+++ react/logisticsInfo.js
@@ -13,6 +13,7 @@ import {
 } from './utils'
 
 export function getNewLogisticsInfoIfPickup({
+  activePickupPoint,
   actionAddress,
   actionSearchAddress,
   channel,
@@ -31,7 +32,7 @@ export function getNewLogisticsInfoIfPickup({
 
   const hasFirstPickupSla =
     firstPickupSla &&
-    logisticsInfo.slas.some(sla => sla.id === firstPickupSla.id)
+    logisticsInfo.slas.some((sla) => sla.id === firstPickupSla.id)
 
   if (
     hasFirstPickupSla &&
@@ -50,7 +51,7 @@ export function getNewLogisticsInfoIfPickup({
 
   const defaultSlaSelection =
     logisticsInfo.slas.find(
-      sla => firstPickupSla && sla.id === firstPickupSla.id
+      (sla) => firstPickupSla && sla.id === firstPickupSla.id
     ) || findSlaWithChannel(logisticsInfo, channel)
 
   if (
@@ -62,16 +63,30 @@ export function getNewLogisticsInfoIfPickup({
       !defaultSlaSelection)
   ) {
     const defaultDeliverySla = findSlaWithChannel(logisticsInfo, DELIVERY)
+
+    if (defaultDeliverySla && hasMultipleItems) {
+      return {
+        ...logisticsInfo,
+        addressId: actionAddress.addressId,
+        selectedDeliveryChannel: DELIVERY,
+        selectedSla: defaultDeliverySla.id,
+      }
+    }
+
+    if (activePickupPoint && activePickupPoint.id) {
+      return {
+        ...logisticsInfo,
+        addressId: actionSearchAddress.addressId,
+        selectedDeliveryChannel: channel,
+        selectedSla: activePickupPoint.id,
+      }
+    }
+
     return {
       ...logisticsInfo,
-      addressId:
-        defaultDeliverySla && hasMultipleItems
-          ? actionAddress.addressId
-          : actionSearchAddress.addressId,
-      selectedDeliveryChannel:
-        defaultDeliverySla && hasMultipleItems ? DELIVERY : channel,
-      selectedSla:
-        defaultDeliverySla && hasMultipleItems ? defaultDeliverySla.id : null,
+      addressId: actionSearchAddress.addressId,
+      selectedDeliveryChannel: channel,
+      selectedSla: null,
     }
   }
 
@@ -132,7 +147,7 @@ export function getNewLogisticsInfoIfDelivery({
   if (
     hasCurrentDeliveryChannel(logisticsInfo, channel) &&
     slaFromSlaOption &&
-    logisticsInfo.slas.some(sla => sla.id === slaFromSlaOption.id)
+    logisticsInfo.slas.some((sla) => sla.id === slaFromSlaOption.id)
   ) {
     return {
       ...logisticsInfo,
@@ -171,7 +186,7 @@ export function getNewLogisticsInfo({
   const actionAddress = removeAddressValidation(action.address)
   const actionSearchAddress = removeAddressValidation(action.searchAddress)
 
-  return newLogisticsInfo.map(li => {
+  return newLogisticsInfo.map((li) => {
     const hasSellerIdMatch = isFromCurrentSeller({ items, li, seller })
 
     if (!hasSellerIdMatch) {
@@ -180,6 +195,7 @@ export function getNewLogisticsInfo({
 
     if (isPickup(channel)) {
       const newLogisticsInfoIfPickup = getNewLogisticsInfoIfPickup({
+        activePickupPoint: action.activePickupPoint,
         action,
         actionAddress,
         actionSearchAddress,